requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==3.2.8
django-tinymce4-lite==1.8.0
gunicorn==20.1.0
jsmin==3.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
wincertstore==0.2
